pyproject.toml
[tool.poetry]
name = "backend-assignment"
version = "0.1.0"
description = ""
authors = ["Kwon Gi Hwan <rlghks786@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.101.1"
sqlalchemy = "^2.0.20"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
alembic = "^1.11.3"
passlib = "^1.7.4"
mysqlclient = "^2.2.0"
python-multipart = "^0.0.6"
uvicorn = "^0.23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
